feat(Prefixes): Moves prefixes directly under src.
